requirements.txt
requests
ipython
Django
djangorestframework
djangorestframework-authtoken